Removing unconsolidated step iterator
